Update README.md
This project was completed as part of my internship/mentorship at Avetti, a SaaS-based company that provides scalable e-commerce solutions. I designed and developed the complete UI/UX for a furniture e-commerce website, which demonstrates the application of Avetti’s SaaS platform. The project includes wireframes, prototypes, and a deployed responsive website